[config] Add single-ads.html file
This is first step to add google ads where I want.

Default layout is used 'single' and it is from here.
https://github.com/mmistakes/minimal-mistakes/blob/master/docs/_layouts/single.html

Copy and paste it to single-ads.html
and use it as default layout.
With this patch, I can customize default layout.
I can use overriding way with saving it as single.html
but to distinguish with original layout, used single-ads.html

Signed-off-by: Jaemin Seo <86sjmjm@gmail.com>